New layer added to neural network framework
The layer is an ansatz layer, where all ansatzes from the VQE framework can be used as neural network layer. This provides the possibility of an exponential decrease in the amount of layers in a feed forward neural network. main_QROTNN.py showcases that two layers with minimal nodes are able to get above 90% accuracy on the iris dataset.